+++ ThreeDSecure/src/main/java/com/braintreepayments/api/ThreeDSecureRequest.java
@@ -40,7 +40,7 @@ public class ThreeDSecureRequest implements Parcelable {
     private boolean challengeRequested = false;
     private boolean dataOnlyRequested = false;
     private boolean exemptionRequested = false;
-    private boolean cardAdd = false;
+    private boolean cardAddChallengeRequested = false;
     private ThreeDSecureV2UiCustomization v2UiCustomization;
     private ThreeDSecureV1UiCustomization v1UiCustomization;
 
@@ -167,11 +167,11 @@ public class ThreeDSecureRequest implements Parcelable {
      * Optional. An authentication created using this flag should only be used for adding a
      * payment method to the merchant's vault and not for creating transactions.
      *
-     * @param cardAdd If set to true, the authentication challenge will be requested from the issuer
+     * @param cardAddChallengeRequested If set to true, the authentication challenge will be requested from the issuer
      *                to confirm adding new card to the merchant's vault.
      */
-    public void setCardAdd(boolean cardAdd) {
-        this.cardAdd = cardAdd;
+    public void setCardAddChallengeRequested(boolean cardAddChallengeRequested) {
+        this.cardAddChallengeRequested = cardAddChallengeRequested;
     }
 
     /**
@@ -288,8 +288,8 @@ public class ThreeDSecureRequest implements Parcelable {
      * @return If the authentication challenge will be requested from the issuer to confirm adding
      * new card to the merchant's vault.
      */
-    public boolean isCardAdd() {
-        return cardAdd;
+    public boolean isCardAddChallengeRequested() {
+        return cardAddChallengeRequested;
     }
 
     /**
@@ -328,7 +328,7 @@ public class ThreeDSecureRequest implements Parcelable {
         dest.writeByte(challengeRequested ? (byte) 1 : 0);
         dest.writeByte(dataOnlyRequested ? (byte) 1 : 0);
         dest.writeByte(exemptionRequested ? (byte) 1 : 0);
-        dest.writeByte(cardAdd ? (byte) 1 : 0);
+        dest.writeByte(cardAddChallengeRequested ? (byte) 1 : 0);
         dest.writeParcelable(v2UiCustomization, flags);
         dest.writeParcelable(v1UiCustomization, flags);
         dest.writeString(accountType);
@@ -346,7 +346,7 @@ public class ThreeDSecureRequest implements Parcelable {
         challengeRequested = in.readByte() > 0;
         dataOnlyRequested = in.readByte() > 0;
         exemptionRequested = in.readByte() > 0;
-        cardAdd = in.readByte() > 0;
+        cardAddChallengeRequested = in.readByte() > 0;
         v2UiCustomization = in.readParcelable(ThreeDSecureV2UiCustomization.class.getClassLoader());
         v1UiCustomization = in.readParcelable(ThreeDSecureV1UiCustomization.class.getClassLoader());
         accountType = in.readString();
@@ -381,8 +381,8 @@ public class ThreeDSecureRequest implements Parcelable {
             base.put("additional_info", additionalInfo);
             base.putOpt("account_type", accountType);
 
-            if (cardAdd) {
-               base.put("card_add", cardAdd);
+            if (cardAddChallengeRequested) {
+               base.put("card_add", cardAddChallengeRequested);
             }
 
             additionalInfo.putOpt("mobile_phone_number", getMobilePhoneNumber());